requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
exceptiongroup==1.3.1
Faker==37.12.0
idna==3.11
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
